refactor: favorites module.
