+++ packages/ui/src/home.tsx
@@ -12,7 +12,7 @@ import { useTheme } from './contexts/settings/theme.js'
 import { BusSelector } from './components/busSelector.js'
 import { Loading } from './components/loading.js'
 import { Settings } from './components/settings/index.js'
-import { Favorites } from './components/favorites.js'
+import { Favorites } from './modules/favorites/components/favorites.js'
 import { Info } from './components/info.js'
 import { Predictions } from './components/predictions.js'
 import { Anchor } from './components/anchor.js'
