+++ packages/ui/src/modules/favorites/components/favorites.tsx
@@ -1,4 +1,4 @@
-import { useStorage } from '../contexts/storage.js'
+import { useStorage } from '../../../contexts/storage.js'
 
 import type { FC } from 'react'
