+++ packages/ui/src/modules/favorites/types.ts
@@ -1,0 +1,10 @@
+import type { Agency, RouteName, DirectionName, Stop } from '../../types.js'
+
+interface Favorite {
+  agency: Agency
+  route: RouteName
+  direction: DirectionName
+  stop: Stop
+}
+
+export type { Favorite }
