+++ packages/ui/tsconfig.json
@@ -4,7 +4,11 @@
     "module": "ESNext",
     "moduleResolution": "Bundler",
     "jsx": "react-jsx",
-    "types": ["vite/client"]
+    "types": ["vite/client"],
+    "paths": {
+      "@core/*": ["./src/*"],
+      "@module/*": ["./src/modules/*"]
+    }
   },
   "include": ["vite.config.ts", "src"]
 }
