+++ packages/ui/vite.config.ts
@@ -1,4 +1,5 @@
 import { env } from 'node:process'
+import { resolve } from 'node:path'
 
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
@@ -34,6 +35,12 @@ export default defineConfig(() => {
     server: {
       proxy,
       host: true
+    },
+    resolve: {
+      alias: {
+        '@core': resolve('src'),
+        '@module': resolve('src/modules')
+      }
     },
     preview: { proxy }
   }
